+++ src/main/java/com/app/pickcourse/repository/PlanDAO.java
@@ -1,5 +1,6 @@
 package com.app.pickcourse.repository;
 
+import com.app.pickcourse.domain.dto.CourseDTO;
 import com.app.pickcourse.domain.dto.PlanDTO;
 import com.app.pickcourse.domain.dto.RecruitPlanDTO;
 import com.app.pickcourse.domain.dto.TourListDTO;
@@ -69,5 +70,4 @@ public class PlanDAO {
     public List<RecruitPlanDTO> findByMemberIdWithLimit(Long memberId, int offset, int limit) {
         return planMapper.selectMyRecruitPlansWithLimit(memberId, offset, limit);
     }
-
 }
